refactor: перенести приложение blog наружу пакета проекта
django-admin startapp кладёт приложение рядом с пакетом проекта, а не внутрь
него, поэтому python_django_orm_blog/blog переехал в blog.

Метка приложения не изменилась: django берёт её из последнего компонента name,
поэтому миграции остаются валидными, makemigrations --check это подтверждает.

Раскладку показывают уроки python_django_orm_course, они правятся отдельными
коммитами в репозиториях курсов ru и es.
